refactor: move parsing code to parser.py
